Machine_learning_pipeline/requirements.txt
# Environment
python-dotenv==1.0.0

# Database drivers
pymysql==1.1.0
pymongo==4.6.0

# Data processing
pandas==2.1.3
